+++ SimulationEngineLib/law/gravity/helper/GpuNewtonGravityHelper.cu
@@ -7,7 +7,7 @@
 #include <cmath>
 
 __device__ 
-void radiusComponentKernelHelper(int idx, Particle** particles, Vector3D* accelerations, int betweenParticlesTriangularCount, double G, int vectorsProcessedTriangular) {
+void radiusComponentKernelHelper(unsigned long long idx, Particle** particles, Vector3D* accelerations, unsigned long long betweenParticlesTriangularCount, double G, unsigned long long vectorsProcessedTriangular) {
 	unsigned long long x, y;
 	MatrixMaths::getLowerTriangularCoordinates(idx + vectorsProcessedTriangular, &x, &y);
 	Vector3D devicePRadiusComponent = getRadiusComponent(particles[x], particles[y], G);
@@ -16,8 +16,8 @@ void radiusComponentKernelHelper(int idx, Particle** particles, Vector3D* accele
 }
 
 __device__ 
-void addAccelerationsKernelLowerHelper(int idx, Particle** particles, Vector3D* accelerations, int x0, int y, int n, int vectorsProcessedTriangular) {
-	int x = idx + x0;
+void addAccelerationsKernelLowerHelper(unsigned long long idx, Particle** particles, Vector3D* accelerations, unsigned long long x0, unsigned long long y, unsigned long long n, unsigned long long vectorsProcessedTriangular) {
+	unsigned long long x = idx + x0;
 	if(x < n) { 
 		unsigned long long radiusComponentIndex = MatrixMaths::getLowerTriangularIndex(x, y);
 		runOnParticle(particles[x], accelerations[radiusComponentIndex - vectorsProcessedTriangular]);
@@ -25,8 +25,8 @@ void addAccelerationsKernelLowerHelper(int idx, Particle** particles, Vector3D* 
 }
 
 __device__ 
-void addAccelerationsKernelUpperHelper(int idx, Particle** particles, Vector3D* accelerations, int x0, int y, int n, int vectorsProcessedTriangular, int particlesProcessed, int betweenParticlesTriangularCount) {
-	int x = idx + x0;
+void addAccelerationsKernelUpperHelper(unsigned long long idx, Particle** particles, Vector3D* accelerations, unsigned long long x0, unsigned long long y, unsigned long long n, unsigned long long vectorsProcessedTriangular, unsigned long long particlesProcessed, unsigned long long betweenParticlesTriangularCount) {
+	unsigned long long x = idx + x0;
 	if(x < n && x >= particlesProcessed) { 
 		unsigned long long radiusComponentIndex = MatrixMaths::getUpperTriangularIndex(x, y);
 		runOnParticle(particles[x], accelerations[radiusComponentIndex - vectorsProcessedTriangular + betweenParticlesTriangularCount]);
